+++ src/todo_list/task.py
@@ -1,3 +1,5 @@
+from datetime import date
+
 from project import Project
 from schedule import Schedule
 
@@ -16,18 +18,32 @@ class Task:
         self.related_tasks = {}  # this is for future use
         self.goal = ""
         self.habit = False  # this is for future use
+        self.is_active = True
         # determine how long it takes to complete related projects/tasks on average
         # to generate estimate for current project/task
         # self.time_spent = {date: timedelta} # this tracks how long I've spent on this task(s)
 
     def unpack_records(self, record):
-        # questions: how do you enforce one element is one of those types
         for i, key in enumerate(self.__dict__):
-            self.__dict__[key] = record[i]
+            if key == "schedule":
+                schedule = Schedule()
+                schedule.unpack_json(record[i])
+                self.__dict__[key] = schedule
+            else:
+                self.__dict__[key] = record[i]
 
     def update(self, task):
         for i, key in enumerate(self.__dict__):
             if task.__dict__[key] is not None:
                 self.__dict__[key] = task.__dict__[key]
-        # todo
         return None
+
+    def update_next_occurrence(self):  # takes a task and then changes the next occurrence date
+        if self.next_occurrence not in self.is_complete:
+            self.next_occurrence = date.today()
+        else:
+            self.next_occurrence = self.schedule.get_next_occurrence()
+
+    def add_next_occurrence(self):
+        self.next_occurrence = self.schedule.get_next_occurrence()
+
